Change to share link with hash is included
Why: earlier, if a user goes to a general page then click on a side menu
link, and then choose to share to socialMedia using the button on the
footer, the hash (in-page link) is not included. This is because the
shared URL is set from the page loading and not after an in-page link is
click.
This commit tried to fix that. It seems to work for Twitter. However,
Facebook and LinkedIn still seem to not work.
This commit also open the share link in another window or tab and not use
pop-up window (because it is usually frowned upon for poor accessibility)