Add: Search and scroll to found value in side menu with dropdown index
Why: for pages with side menu and dropdown index, sometimes, search
function is neccessary. This commit add a search function for a tag
using the href value. If search term is found, the sidebar will scroll
to the first item
Some properties of this search includes:
- Case INSENSITIVE search
- Work best with single word search because it uses href contains value
- It will only sroll to the first match
- It will highlight the found search term
- If term is not found, a pop up will display. This popup will be closed
automatically in 5 second or after a successful search, whatever comes
first.